qwen_function/function/requirements.txt
requests
boto3
boto3-stubs[essential]
llama-cpp-python==0.2.90